Implement a CLI using docopt
Uses the old app class in a hacky way to make the CLI work.
